FileLoaderDemo class added; Logging implemented
