+++ task5/src/main/resources/linksToDownload.properties
@@ -4,3 +4,4 @@ https://www.kmu.gov.ua/storage/app/sites/1/vnutrishnii%20audyt/deklaratsiia-vnut
 https://www.kmu.gov.ua/storage/app/sites/1/konkursy/arhiv/ofis-reform/informatsiya-pro-konkursniy-vidbir-do-23102017.docx
 https://www.seco-cooperation.admin.ch/secocoop/en/home.html
 https://cdn.webshopapp.com/shops/94414/files/52440074/flag-of-ukraine.jpg
+https://learnenglish.britishcouncil.org/sites/podcasts/files/learnenglish-big-city-small-world-01-02-support-pack.pdf
